+++ src/lib.rs
@@ -6,7 +6,7 @@
 use crate::action_state::ActionState;
 use crate::input_map::InputMap;
 use bevy::ecs::prelude::*;
-use bevy::reflect::Reflect;
+use bevy::reflect::{FromReflect, Reflect, TypePath};
 use std::hash::Hash;
 use std::marker::PhantomData;
 
@@ -83,7 +83,9 @@ pub mod prelude {
 ///    Ultimate,
 /// }
 /// ```
-pub trait Actionlike: Eq + Hash + Send + Sync + Clone + Hash + Reflect + 'static {
+pub trait Actionlike:
+    Eq + Hash + Send + Sync + Clone + Hash + Reflect + TypePath + FromReflect + 'static
+{
     /// The number of variants of this action type
     fn n_variants() -> usize;