+++ RELEASES.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Removed `multimap` dependency in favor of regular `HashMap` which allowed to derive `Reflect` for `InputMap`.
+- Register types in the reflection system.
 - added support in `ActionDiff` for value and axis_pair changes
 - Added `InputMap::Clear`.